requirements.txt
django>=5.0
psycopg2-binary
dj-database-url
djangorestframework
python-dotenv==1.0.0
social-auth-app-django
djangorestframework-simplejwt
PyJWT
requests